+++ CHANGES.rst
@@ -1,8 +1,8 @@
 v4.1.0
 ======
 
-``Projection`` now accepts an iterable or callable for its
-first argument.
+``Projection`` now accepts an iterable or callable or pattern
+for matching keys.
 
 v4.0.0
 ======